requirements_openai.txt
langchain==0.0.154
streamlit==1.18.1
PyPDF2==3.0.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
streamlit-extras
altair==4.0
tiktoken
openai==0.28.1
fpdf==1.7.2
googletrans==3.1.0a0
httpcore==0.13.3
numpy==1.21.5
opencv_contrib_python==4.8.0.76
opencv_python==4.9.0.80
opencv_python_headless==4.8.0.74
Pillow==10.1.0
pydub==0.25.1
reportlab==3.6.8
Requests==2.31.0
SpeechRecognition==3.10.1
websocket_client==1.6.4
psycopg2-binary==2.9.9
scipy==1.12.0
termcolor==2.4.0
graphviz==0.20.3
selenium==4.1.0
#-e .